pyproject.toml
[project]
name = "food-ordering-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "anthropic>=0.46.0",
    "mcp>=1.2.1",
    "nest-asyncio>=1.6.0",
    "permit>=2.7.2",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
]
